Building HTML with an AJAX Call
This AJAX call builds out all of the HTML based on a dynamic "eventLevel"  being passed through the the class ".eventLevel". Based on that event level the HTML "circle" class is changed to reflect the "if else" statement. All of the html is built dynamically after the "if else" statement based on database queries.